lab_files/requirements.txt
chainlit
langchain
langchain_openai
python-dotenv
mssql-python
azure-identity